+++ src/gateway/common/constants.py
@@ -1,3 +1,8 @@
+# RabbitMQ constants
+LIVE_EVENTS_EXCHANGE_NAME = "liveEventsExchange"
+LIVE_EVENTS_QUEUE_NAME = "liveEventsQueue"
+
+# Other MS Addresses
 TRIP_RESEARCHER_SERVICE_ADDRESS = "http://localhost:2137"
 RESERVATION_MS_ADDRESS = "http://localhost:2138"
 PAYMENT_MS_ADDRESS = "http://localhost:2139"
