+++ src/gateway/endpoints/events.py
@@ -3,6 +3,7 @@ import json
 from fastapi import APIRouter, status
 from starlette.responses import JSONResponse, Response
 
+from common.constants import LIVE_EVENTS_QUEUE_NAME, LIVE_EVENTS_EXCHANGE_NAME
 from rabbitmq.live_events import LiveEvents
 from rabbitmq.rabbitmq_client import RabbitMQClient
 
@@ -45,7 +46,8 @@ async def publish_event():
             "transport_type": "rower"
         }]
     }
-    RabbitMQClient.send_data_to_queue(queue_name="liveEventsQueue",
-                                      payload=json.dumps(example_payload, ensure_ascii=False).encode('utf-8'))
+    RabbitMQClient.send_data_to_queue(queue_name=LIVE_EVENTS_QUEUE_NAME,
+                                      payload=json.dumps(example_payload, ensure_ascii=False).encode('utf-8'),
+                                      exchange_name=LIVE_EVENTS_EXCHANGE_NAME)
     return Response(status_code=status.HTTP_201_CREATED, content=f"elo",
                     media_type="text/plain")
