+++ src/gateway/endpoints/purchases.py
@@ -2,8 +2,8 @@ import requests
 from fastapi import APIRouter, Response, Depends, status
 from starlette.responses import JSONResponse
 
-from common.addresses import PURCHASE_MS_ADDRESS
 from common.authentication import oauth2_scheme, verify_jwt_token
+from common.constants import PURCHASE_MS_ADDRESS
 from users.service import verify_user_identify
 
 router = APIRouter(prefix="/api/v1/purchases")
