+++ src/gateway/main.py
@@ -4,12 +4,14 @@ import threading
 import uvicorn
 from fastapi import FastAPI, APIRouter
 
+from common.constants import LIVE_EVENTS_QUEUE_NAME
 from endpoints.events import router as event_router
 from endpoints.payments import router as payments_router
 from endpoints.purchases import router as purchases_router
 from endpoints.reservations import router as reservations_router
 from endpoints.trips import router as trips_router
 from endpoints.users import router as users_router
+from rabbitmq.live_events import consume_live_event
 from rabbitmq.rabbitmq_client import RabbitMQClient
 
 app = FastAPI()
@@ -28,6 +30,9 @@ if __name__ == "__main__":
     loop = asyncio.new_event_loop()
     asyncio.set_event_loop(loop)
     thread = threading.Thread(target=loop.run_until_complete,
-                              args=(RabbitMQClient.create_connection_and_start_consuming(),))
+                              args=(
+                                  RabbitMQClient.create_connection_and_start_consuming(
+                                      queue_name=LIVE_EVENTS_QUEUE_NAME,
+                                      consume_function=consume_live_event),))
     thread.start()
     uvicorn.run(app, host="127.0.0.1", port=8000)
